feat : VoteDetailViewController 추가 & 기존 투표 상세 화면 관련 코드 수정
- 이중 컬렉션뷰를 사용하지 않는 투표 상세 화면을 만들기 위해 VoteDetailViewController 추가
- 기존 VoteDetailViewController -> VoteDetailTableViewController로 수정 후 스토리보드와 연결 해제
